+++ src/Client/ClientCore/Systems/ClientState/ClientStateServices.cs
@@ -1,0 +1,35 @@
+// Sovereign Engine
+// Copyright (c) 2024 opticfluorine
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+namespace Sovereign.ClientCore.Systems.ClientState;
+
+/// <summary>
+///     Exports public services from the ClientState system to the rest of the client.
+/// </summary>
+public class ClientStateServices
+{
+    private readonly ClientStateMachine stateMachine;
+
+    public ClientStateServices(ClientStateMachine stateMachine)
+    {
+        this.stateMachine = stateMachine;
+    }
+
+    /// <summary>
+    ///     Current state of the top-level client state machine.
+    /// </summary>
+    public ClientState State => stateMachine.State;
+}